+++ src/countries/fr.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const fr: Country = {
+	code: 'fr',
+	name: {
+		en: 'France',
+		ar: 'فرنسا',
+		de: 'Frankreich',
+		sv: 'Frankrike',
+		da: 'Frankrig',
+		sq: 'Franca',
+		pt: 'França',
+		tr: 'Fransa',
+		fa: 'فرانسه',
+		uk: 'Франція',
+		xa: 'La Marseillaise',
+		xt: 'La Marseillaise',
+	},
+	flag: '🇫🇷',
+	beta: true,
+}