+++ src/countries/no.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const no: Country = {
+	code: 'no',
+	name: {
+		en: 'Norway',
+		ar: 'النرويج',
+		de: 'Norwegen',
+		sv: 'Norge',
+		da: 'Norge',
+		sq: 'Norvegjia',
+		pt: 'Noruega',
+		tr: 'Norveç',
+		fa: 'نروژ',
+		uk: 'Норвегія',
+		xa: 'Ja, vi elsker dette landet',
+		xt: 'Ja, vi elsker dette landet',
+	},
+	flag: '🇳🇴',
+	beta: true,
+}